% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appendGmtList.R
\name{appendGmtList}
\alias{appendGmtList}
\title{Append a GmtList object to another one}
\usage{
appendGmtList(gmtList, newGmtList, ...)
}
\arguments{
\item{gmtList}{A \code{GmtList} object}

\item{newGmtList}{Another \code{GmtList} object to be appended}

\item{...}{Further \code{GmtList} object to be appended}
}
\value{
A new \code{GmtList} list, with all elements in the input appended in the given order
}
\description{
Append a GmtList object to another one
}
\examples{
test_gmt_file<- system.file("extdata/test.gmt", package="BioQC")
testGmtList1 <- readGmt(test_gmt_file, namespace="test1")
testGmtList2 <- readGmt(test_gmt_file, namespace="test2")
testGmtList3 <- readGmt(test_gmt_file, namespace="test3")
testGmtAppended <- appendGmtList(testGmtList1, testGmtList2, testGmtList3)
}
