% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{BaseIndexList-class}
\alias{BaseIndexList-class}
\title{An S4 class to hold a list of indices, with the possibility to specify the 
offset of the indices. IndexList and SignedIndexList extend this class}
\description{
An S4 class to hold a list of indices, with the possibility to specify the 
offset of the indices. IndexList and SignedIndexList extend this class
}
\section{Slots}{

\describe{
\item{\code{offset}}{An integer specifying the value of first element. Default 1}

\item{\code{keepNA}}{Logical, whether NA is kept during construction}

\item{\code{keepDup}}{Logical, whether duplicated values are kept during construction}
}}

