% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GmtList-funcs.R
\name{filterBySize}
\alias{filterBySize}
\title{Filter a GmtList by size}
\usage{
filterBySize(x, min, max)
}
\arguments{
\item{x}{A \code{GmtList} object}

\item{min}{Numeric, gene-sets with fewer genes than \code{min} will be removed}

\item{max}{Numeric, gene-sets with more genes than \code{max} will be removed}
}
\value{
A \code{GmtList} object with sizes (count of genes) between \code{min} and \code{max} (inclusive).
}
\description{
Filter a GmtList by size
}
