% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{GmtList}
\alias{GmtList}
\title{Convert a list to a GmtList object}
\usage{
GmtList(list)
}
\arguments{
\item{list}{A list of genesets; each geneset is a list of at least three fields: 'name', 'desc', and 'genes'. 'name' and 'desc' contains one character string ('desc' can be NULL while 'name' cannot), and 'genes' can be either NULL or a character vector.}
}
\description{
Convert a list to a GmtList object
}
\examples{
testList <- list(list(name="GS_A", desc=NULL, genes=LETTERS[1:3]),
                 list(name="GS_B", desc="gene set B", genes=LETTERS[1:5]),
                 list(name="GS_C", desc="gene set C", genes=NULL))
testGmt <- GmtList(testList)
}
\seealso{
If a list of gene symbols need to be converted into a GmtList, use 'as.gmtlist' instead
}

