% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{gsCategory}
\alias{gsCategory}
\title{Gene-set categories}
\usage{
gsCategory(x)
}
\arguments{
\item{x}{A \code{GmtList} object}
}
\value{
Categories as a vector of character strings of the same length as \code{x}
}
\description{
Gene-set categories
}
