% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GmtList-funcs.R
\name{gsDesc}
\alias{gsDesc}
\title{Gene-set descriptions}
\usage{
gsDesc(x)
}
\arguments{
\item{x}{A \code{GmtList} object}
}
\value{
Descriptions as a vector of character strings of the same length as \code{x}
}
\description{
Gene-set descriptions
}
