% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GmtList-funcs.R
\name{gsGeneCount}
\alias{gsGeneCount}
\alias{gsSize}
\title{Gene-set gene counts}
\usage{
gsGeneCount(x, uniqGenes = TRUE)

gsSize(x, uniqGenes = TRUE)
}
\arguments{
\item{x}{A \code{GmtList} or similar object}

\item{uniqGenes}{Logical, whether only unique genes are counted}
}
\value{
Gene counts (aka gene-set sizes) as a vector of integer of the same length as \code{x}
}
\description{
Gene-set gene counts

gsSize is the synonym of gsGeneCount
}
