% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GmtList-funcs.R
\name{gsGenes}
\alias{gsGenes}
\title{Gene-set member genes}
\usage{
gsGenes(x)
}
\arguments{
\item{x}{A \code{GmtList} object}
}
\value{
A list of genes as character strings of the same length as \code{x}
}
\description{
Gene-set member genes
}
