% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{gsName}
\alias{gsName}
\title{Gene-set names}
\usage{
gsName(x)
}
\arguments{
\item{x}{A \code{GmtList} object}
}
\value{
Names as a vector of character strings of the same length as \code{x}
}
\description{
Gene-set names
}
