% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GmtList-funcs.R
\name{gsNamespace}
\alias{gsNamespace}
\title{Gene-set namespaces}
\usage{
gsNamespace(x)
}
\arguments{
\item{x}{A \code{GmtList} object}
}
\value{
Namespaces as a vector of character strings of the same length as \code{x}
}
\description{
Gene-set namespaces
}
