% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GmtList-funcs.R
\name{hasNamespace}
\alias{hasNamespace}
\title{Whether namespace is set}
\usage{
hasNamespace(x)
}
\arguments{
\item{x}{A \code{GmtList} object}
}
\value{
Logical, whether all gene-sets have the field \code{namespace} set
}
\description{
Whether namespace is set
}
