% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\docType{methods}
\name{IndexList,list-method}
\alias{IndexList,list-method}
\title{Convert a list of numeric (integer) vectors into an index list}
\usage{
\S4method{IndexList}{list}(object, keepNA = FALSE, keepDup = FALSE,
  offset = 1L)
}
\arguments{
\item{object}{A list of numeric (integer) vectors}

\item{offset}{offset; 1 if missing}
}
\description{
Convert a list of numeric (integer) vectors into an index list
}
\examples{
IndexList(list(A=1:3, B=4:5, C=7:9))
IndexList(list(A=1:3, B=4:5, C=7:9), offset=0)
}
