% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\docType{methods}
\name{IndexList,logical-method}
\alias{IndexList,logical-method}
\title{Convert several logical vectors into an index list}
\usage{
\S4method{IndexList}{logical}(object, ..., keepNA = FALSE, keepDup = FALSE,
  offset = 1L)
}
\arguments{
\item{object}{A logical vector}

\item{...}{Other logical vector(s) of the same length as \code{object}}

\item{offset}{offset; 1 if missing}
}
\description{
Convert several logical vectors into an index list
}
\examples{
IndexList(c(FALSE, TRUE, TRUE), c(FALSE, FALSE, TRUE), c(TRUE, FALSE, FALSE), offset=0)
}
