% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllMethods.R
\docType{methods}
\name{IndexList,numeric-method}
\alias{IndexList,numeric-method}
\title{Convert several numeric vectors into an index list}
\usage{
\S4method{IndexList}{numeric}(object, ..., keepNA = FALSE, keepDup = FALSE,
  offset = 1L)
}
\arguments{
\item{object}{A integer (numeric) vector}

\item{...}{Other integer (numeric) vector(s)}

\item{offset}{offset; 1 if missing}
}
\description{
Convert several numeric vectors into an index list
}
\examples{
IndexList(1:3, 4:5, 7:9)
IndexList(1:3, 4:5, 7:9, offset=0)
}

