% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{IndexListFromList}
\alias{IndexListFromList}
\title{Convert a list to an IndexList object}
\usage{
IndexListFromList(list, keepNA = FALSE, keepDup = FALSE, offset = 1L)
}
\arguments{
\item{list}{A list of unique integer indices or NULL, or logical vectors of same lengths. NA is discarded.}

\item{keepNA}{Logical, whether NA indices should be kept or not. Default: FALSE (removed)}

\item{keepDup}{Logical, whether duplicated indices should be kept or not. Default: FALSE (removed)}

\item{offset}{Integer, the starting index. Default: 1 (as in the convention of R)}
}
\description{
Convert a list to an IndexList object
}
\examples{
testList <- list(GS_A=c(1,2,3,4,3),
                 GS_B=c(2,3,4,5),
                 GS_C=NULL,
                 GS_D=c(1,3,5,NA),
                 GS_E=c(2,4))
testIndexList <- IndexList(testList)
}

