% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{isValidBaseIndexList}
\alias{isValidBaseIndexList}
\title{Function to validate a BaseIndexList  object}
\usage{
isValidBaseIndexList(object)
}
\arguments{
\item{object}{A BaseIndexList  object
Use \code{setValidity("BaseIndexList", "isValidBaseIndexList")} to check integrity of BaseIndexList  objects.
It can be very slow, therefore the feature is not turned on by default}
}
\description{
Function to validate a BaseIndexList  object
}
