% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{isValidGmtList}
\alias{isValidGmtList}
\title{Function to validate a GmtList object}
\usage{
isValidGmtList(object)
}
\arguments{
\item{object}{A GmtList object
Use \code{setValidity("GmtList", "isValidGmtList")} to check integrity of GmtList objects.
It can be very slow, therefore the feature is not turned on by default}
}
\description{
Function to validate a GmtList object
}
