% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{isValidIndexList}
\alias{isValidIndexList}
\title{Function to validate an IndexList object}
\usage{
isValidIndexList(object)
}
\arguments{
\item{object}{an IndexList object
Use \code{setValidity("BaseIndexList", "isValidBaseIndexList")} to check integrity of IndexList objects.
It can be very slow, therefore the feature is not turned on by default}
}
\description{
Function to validate an IndexList object
}
