% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{isValidSignedGenesets}
\alias{isValidSignedGenesets}
\title{Function to validate a SignedGenesets object}
\usage{
isValidSignedGenesets(object)
}
\arguments{
\item{object}{A SignedGenesets object
Use \code{setValidity("SignedGenesets", "isValidSignedGenesets")} to check integrity of SignedGenesets objects.
It can be very slow, therefore the feature is not turned on by default}
}
\description{
Function to validate a SignedGenesets object
}
