% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{isValidSignedIndexList}
\alias{isValidSignedIndexList}
\title{Function to validate a SignedIndexList object}
\usage{
isValidSignedIndexList(object)
}
\arguments{
\item{object}{a SignedIndexList object
Use \code{setValidity("SignedIndexList", "isValidSignedIndexList")} to check integrity of SignedIndexList objects.
It can be very slow, therefore the feature is not turned on by default}
}
\description{
Function to validate a SignedIndexList object
}
