% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmwTest.R
\name{matchGenes.default}
\alias{matchGenes.default}
\title{Match genes in a GmtList to a vector of genesymbols}
\usage{
matchGenes.default(gmtList, geneSymbols)
}
\arguments{
\item{gmtList}{A GmtList object}

\item{geneSymbols}{Gene symbols to be matched; they can come from a column in an eSet object, for example}
}
\description{
Match genes in a GmtList to a vector of genesymbols
}

