% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchGenes.R
\name{matchGenes.signedDefault}
\alias{matchGenes.signedDefault}
\title{Match genes in a SignedGenesets to a vector of genesymbols}
\usage{
matchGenes.signedDefault(signedGenesets, geneSymbols)
}
\arguments{
\item{geneSymbols}{Gene symbols to be matched; they can come from a column in an eSet object, for example}

\item{gmtList}{A SignedGenesets object}
}
\description{
Match genes in a SignedGenesets to a vector of genesymbols
}
