% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllMethods.R
\docType{methods}
\name{offset,IndexList-method}
\alias{offset,IndexList-method}
\title{Get offset from an IndexList object}
\usage{
\S4method{offset}{IndexList}(object)
}
\arguments{
\item{object}{An IndexList object}
}
\description{
Get offset from an IndexList object
}
\examples{
myIndexList <- IndexList(list(1:5, 2:7, 3:8), offset=1L)
offset(myIndexList)
}

