% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettySigNames.R
\name{prettySigNames}
\alias{prettySigNames}
\title{Prettify default signature names}
\usage{
prettySigNames(names, includeNamespace = TRUE)
}
\arguments{
\item{names}{Character strings, signature names}

\item{includeNamespace}{Logical, whether the namespace of the signatures should be included}
}
\value{
Character strings, pretty signature names
}
\description{
Prettify default signature names
}
\examples{
sig <- readCurrentSignatures()
prettyNames <- prettySigNames(names(sig))
}
