% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGmt.R
\name{readGmt}
\alias{readGmt}
\title{Read in gene-sets from a GMT file}
\usage{
readGmt(filename, uniqGenes = TRUE)
}
\arguments{
\item{filename}{Character, GMT file name}

\item{uniqGenes}{Logical, whether duplicated genes should be removed}
}
\value{
A \code{GmtList} object, which is a S3-class wrapper of a list. Each 
element in the object is a list of (at least) three items: 
\itemize{
  \item gene-set name (field \code{name}), character string
  \item gene-set description (field \code{desc}), character string
  \item genes (field \code{genes}), a vector of character strings
}
}
\description{
Read in gene-sets from a GMT file
}
\examples{
gmt_file <- system.file("extdata/exp.tissuemark.affy.roche.symbols.gmt", package="BioQC")
gmt_list <- readGmt(gmt_file)
gmt_uniqGenes_list <- readGmt(gmt_file, uniqGenes=TRUE)

}
