% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{setCategory}
\alias{setCategory}
\title{Set the category field in each gene-set within a GmtList}
\usage{
setCategory(x, category)
}
\arguments{
\item{x}{A \code{GmtList} object encoding a list of gene-sets}

\item{category}{It can be either a function that applies to a \code{gene-set list} element of the object (for instance \code{function(x) x$desc} to extract description), or a vector of the same length of \code{x}, or in the special case \code{NULL}, which will erase the field category.

Note that using vectors as \code{category} leads to poor performance when the input object has many gene-sets.}
}
\description{
Set the category field in each gene-set within a GmtList
}
\examples{
myGmtList <- GmtList(list(list(name="GeneSet1", desc="Category1", genes=LETTERS[1:3]),
  list(name="GeneSet2", desc="Category1", genes=rep(LETTERS[4:6],2)),
  list(name="GeneSet1", desc="Category1", genes=LETTERS[4:6]),
  list(name="GeneSet3", desc="Category2", genes=LETTERS[1:5])))
hasCategory(myGmtList)
myGmtList2 <- setCategory(myGmtList, category=function(x) x$desc)
gsCategory(myGmtList2)
## the function can provide flexible ways to encode the gene-set category
myGmtList3 <- setCategory(myGmtList, category=function(x) gsub("Category", "C", x$desc))
gsCategory(myGmtList3)
## using vectors
myGmtList4 <- setCategory(myGmtList, category=c("C1", "C1", "C1", "C2"))
gsCategory(myGmtList4)
myGmtList2null <- setCategory(myGmtList2, category=NULL)
hasCategory(myGmtList2null)
}
