% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{setDescAsCategory}
\alias{setDescAsCategory}
\title{Set gene-set description as category}
\usage{
setDescAsCategory(x)
}
\arguments{
\item{x}{A \code{GmtList} object

This function wrapps \code{setCategory} to set gene-set description as category}
}
\description{
Set gene-set description as category
}
\seealso{
\code{\link{setCategory}}
}
