% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{setDescAsNamespace}
\alias{setDescAsNamespace}
\title{Set gene-set description as namespace}
\usage{
setDescAsNamespace(x)
}
\arguments{
\item{x}{A \code{GmtList} object

This function wrapps \code{setNamespace} to set gene-set description as namespace}
}
\description{
Set gene-set description as namespace
}
\seealso{
\code{\link{setNamespace}}
}
