% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{gsCategory<-}
\alias{gsCategory<-}
\title{Set the category field in each gene-set within a GmtList}
\usage{
gsCategory(x) <- value
}
\arguments{
\item{x}{A \code{GmtList} object}

\item{value}{Category, a function, \code{NULL}, or a vector. See \code{setGsCategory}}
}
\description{
Set the category field in each gene-set within a GmtList
}
\seealso{
\code{\link{setGsCategory}}
}
