% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{gsNamespace<-}
\alias{gsNamespace<-}
\title{gsNamespace<- is the synonym of setGsNamespace}
\usage{
gsNamespace(x) <- value
}
\arguments{
\item{x}{A \code{GmtList} object}

\item{value}{\code{namespace} in \code{setGsNamespace}. It can be either a function that applies to a \code{gene-set list} element of the object (for instance \code{function(x) x$desc} to extract description), or a vector of the same length of \code{x}, or in the special case \code{NULL}, which will erase the field namespace.}
}
\description{
gsNamespace<- is the synonym of setGsNamespace
}
