% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{show,GmtList-method}
\alias{show,GmtList-method}
\title{Show method for GmtList}
\usage{
\S4method{show}{GmtList}(object)
}
\arguments{
\item{object}{An object of the class \code{GmtList}}
}
\description{
Show method for GmtList
}
