% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{show,IndexList-method}
\alias{show,IndexList-method}
\title{Show method for IndexList}
\usage{
\S4method{show}{IndexList}(object)
}
\arguments{
\item{object}{An object of the class \code{IndexList}}
}
\description{
Show method for IndexList
}
