% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\docType{methods}
\name{show,SignedGenesets-method}
\alias{show,SignedGenesets-method}
\title{Show method for SignedGenesets}
\usage{
\S4method{show}{SignedGenesets}(object)
}
\arguments{
\item{object}{An object of the class \code{SignedGenesets}}
}
\description{
Show method for SignedGenesets
}
