% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{SignedGenesets-class}
\alias{SignedGenesets-class}
\title{An S4 class to hold signed genesets, each item in the list is in in turn a 
list containing following items: name, pos, and neg.}
\description{
An S4 class to hold signed genesets, each item in the list is in in turn a 
list containing following items: name, pos, and neg.
}
