% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{SignedIndexList-class}
\alias{SignedIndexList-class}
\title{An S4 class to hold a list of signed integers as indices, with the possibility to specify the offset of the indices}
\description{
An S4 class to hold a list of signed integers as indices, with the possibility to specify the offset of the indices
}
\section{Slots}{

\describe{
\item{\code{offset}}{An integer specifying the value of first element. Default 1}

\item{\code{keepNA}}{Logical, whether NA was kept during construction}

\item{\code{keepDup}}{Logical, whether duplicated values were kept during construction}
}}

