% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{simplifyMatrix}
\alias{simplifyMatrix}
\title{Simplify matrix in case of single row/columns}
\usage{
simplifyMatrix(matrix)
}
\arguments{
\item{matrix}{A matrix of any dimension

If only one row/column is present, the dimension is dropped and a vector will be returned}
}
\description{
Simplify matrix in case of single row/columns
}
\examples{
testMatrix <- matrix(round(rnorm(9),2), nrow=3)
simplifyMatrix(testMatrix)
simplifyMatrix(testMatrix[1L,,drop=FALSE])
simplifyMatrix(testMatrix[,1L,drop=FALSE])
}

