% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{[.GmtList}
\alias{[.GmtList}
\title{Subsetting GmtList object into another GmtList object}
\usage{
\method{[}{GmtList}(x, i, drop = FALSE)
}
\arguments{
\item{x}{A GmtList object}

\item{i}{Index to subset}

\item{drop}{In case only one element remains, should a list representing the single geneset returned? Default: FALSE}
}
\description{
Subsetting GmtList object into another GmtList object
}
\examples{
myGmtList <- GmtList(list(gs1=letters[1:3], gs2=letters[3:4], gs3=letters[4:5]))
myGmtList[1:2]
myGmtList[1] ## default behaviour: not dropping
myGmtList[1,drop=TRUE] ## force dropping
}
