% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{[[.GmtList}
\alias{[[.GmtList}
\title{Subsetting GmtList object to fetch one gene-set}
\usage{
\method{[[}{GmtList}(x, i)
}
\arguments{
\item{x}{A GmtList object}

\item{i}{The index to subset}
}
\description{
Subsetting GmtList object to fetch one gene-set
}
\examples{
myGmtList <- GmtList(list(gs1=letters[1:3], gs2=letters[3:4], gs3=letters[4:5]))
myGmtList[[1]]
}
