% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGmt.R
\name{as.gmtlist}
\alias{as.gmtlist}
\title{Convert a list of gene symbols into a gmtlist}
\usage{
as.gmtlist(list, description = NULL, uniqGenes = TRUE)
}
\arguments{
\item{list}{A named list with character vectors of genes. Names will become names of gene sets; character vectors will become genes}

\item{description}{Character, description of gene-sets. The value will be expanded to the same length of the list.}

\item{uniqGenes}{Logical, whether redundant genes should be made unique?}
}
\description{
Convert a list of gene symbols into a gmtlist
}
\examples{
testVec <- list(GeneSet1=c("AKT1", "AKT2"),
               GeneSet2=c("MAPK1", "MAPK3"),
               GeneSet3=NULL)
testVecGmtlist <- as.gmtlist(testVec)

}
