% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R, R/matchGenes.R
\docType{methods}
\name{matchGenes}
\alias{matchGenes}
\alias{matchGenes}
\alias{matchGenes,GmtList,character-method}
\alias{matchGenes,GmtList,matrix-method}
\alias{matchGenes,GmtList,eSet-method}
\alias{matchGenes,character,character-method}
\alias{matchGenes,character,matrix-method}
\alias{matchGenes,character,eSet-method}
\alias{matchGenes,SignedGenesets,character-method}
\alias{matchGenes,SignedGenesets,matrix-method}
\alias{matchGenes,SignedGenesets,eSet-method}
\title{Match genes in a list-like object to a vector of genesymbols}
\usage{
matchGenes(list, object, ...)

\S4method{matchGenes}{GmtList,character}(list, object)

\S4method{matchGenes}{GmtList,matrix}(list, object)

\S4method{matchGenes}{GmtList,eSet}(list, object, col = "GeneSymbol")

\S4method{matchGenes}{character,character}(list, object)

\S4method{matchGenes}{character,matrix}(list, object)

\S4method{matchGenes}{character,eSet}(list, object)

\S4method{matchGenes}{SignedGenesets,character}(list, object)

\S4method{matchGenes}{SignedGenesets,matrix}(list, object)

\S4method{matchGenes}{SignedGenesets,eSet}(list, object, col = "GeneSymbol")
}
\arguments{
\item{list}{A GmtList, list, character or SignedGenesets object}

\item{object}{Gene symbols to be matched; they can come from a column in an 
eSet object, for example}

\item{...}{additional arguments like col}

\item{col}{A string used with an \code{eSet}, currently defaults to "GeneSymbol"}
}
\description{
Match genes in a list-like object to a vector of genesymbols
}
