% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\docType{methods}
\name{SignedIndexList}
\alias{SignedIndexList}
\alias{SignedIndexList}
\alias{SignedIndexList,list-method}
\title{Convert a list into a SignedIndexList}
\usage{
SignedIndexList(object, ...)

\S4method{SignedIndexList}{list}(object, keepNA = FALSE, keepDup = FALSE,
  offset = 1L)
}
\arguments{
\item{object}{A list of atleast one list of atleast one list or Vector called 
either `pos` or `neg`}

\item{keepNA}{Logical, whether NA indices should be kept or not. Default: 
FALSE (removed)}

\item{keepDup}{Logical, whether duplicated indices should be kept or not. 
Default: FALSE (removed)}

\item{offset}{offset; 1 if missing}
}
\value{
A SignedIndexList of lists (named like the second list-level of the 
input) containing two vectors named `positive` and `negative`, which contain 
the same Argumetns as the IndexList resulting of the `pos` and `neg` lists
or vectors of the input.
}
\description{
Convert a list into a SignedIndexList
}
\examples{
myList <- list(a = list(pos = list(1, 2, 2, 4), neg = c(TRUE, FALSE, TRUE)), 
b = list(NA), c = list(pos = c(c(2, 3), c(1, 3))))
SignedIndexList(myList)
}
