% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wmwTest.R
\name{wmwTestInR}
\alias{wmwTestInR}
\title{Wilcoxon-Mann-Whitney test in R}
\usage{
wmwTestInR(x, sub, valType = c("two.sided", "less", "greater"),
  statistic = FALSE)
}
\arguments{
\item{x}{A numerical vector}

\item{sub}{A logical vector or integer vector to subset \code{x}. Numbers in \code{sub} are compared with numbers out of \code{sub}}

\item{statistic}{Logical, if \code{TRUE}, the U-statistic of the Wilcoxon-Mann-Whitney test is returned; otherwise the p-value is returned}

\item{alternative}{two.sided, less, or greater}
}
\description{
Wilcoxon-Mann-Whitney test in R
}
\examples{
testNums <- 1:10
testSub <- rep_len(c(TRUE, FALSE), length.out=length(testNums))
wmwTestInR(testNums, testSub)
}

